Move PaginationHeaders into `api` package
